Slice: Add a get() to force casting.
An operator. would be really nice here.

This does not allow `a(0).get().replace(0, "foo");` since a new object
is created at get().  There's no sane way to know when to slice the
parent object outside of when it is casted.

    String foo = "01234";
    foo(0).replace(1, "!");

Here I would expect foo to change to "0!234" but the same interface is
ambiguous here where it should not cause a slicing of foo to "!234":

    String foo = "01234";
    foo(1).replace(0, "!");

Here "0!234" is still expected but the act of casting would need to
slice it; there's no way to hook in the final commit without overriding
every member function in the Slice proxy to do the right thing like:

    void replace(size_t rstart, T&& item) {
      rca.replace(start + rstart, std::move(item), len);
    }

    void replace(size_t rstart, T&& item, size_t rlen) {
      rca.replace(start + rstart, std::move(item), len + rlen);
    }